DBI {gamlss.dist}	R Documentation

The Double binomial distribution

Description

The function DBI() defines the double binomial distribution, a two parameters distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dDBI, pDBI, qDBI and rDBI define the density, distribution function, quantile function and random generation for the double binomial, DBI(), distribution. The function GetBI_C calculates numericaly the constant of proportionality needed for the pdf to sum up to 1.

Usage

DBI(mu.link = "logit", sigma.link = "log")
dDBI(x, mu = 0.5, sigma = 1, bd = 2, log = FALSE)
pDBI(q, mu = 0.5, sigma = 1, bd = 2, lower.tail = TRUE, 
       log.p = FALSE)
qDBI(p, mu = 0.5, sigma = 1, bd = 2, lower.tail = TRUE, 
       log.p = FALSE)
rDBI(n, mu = 0.5, sigma = 1, bd = 2)       
GetBI_C(mu, sigma, bd)       

Arguments

mu.link
the link function for mu with default log

sigma.link
the link function for sigma with default log

x, q
vector of (non-negative integer) quantiles

bd
vector of binomial denominator

p
vector of probabilities

mu
the mu parameter

sigma
the sigma parameter

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

log, log.p
logical; if TRUE, probabilities p are given as log(p)

n
how many random values to generate

Details

The definition for the Double Poisson distribution first introduced by Efron (1986) is:

f(y| n, μ,σ)=[1/C(n,μ,σ)] [Γ(n+1)/Γ(y+1)Γ(n-y+1)] [y^y (n-y)^{n-y}/n^n][n^{n/σ} μ^{y/σ} ( 1-μ)^{(n-y)/σ}/ y^{y/σ} ( n-y)^{(n-y)/σ}]

for y=0,1,2, ...,Inf, μ>0 and σ>0 where C is the constant of proportinality which is calculated numerically using the function GetBI_C().

Value

The function DBI returns a gamlss.family object which can be used to fit a double binomial distribution in the gamlss() function.

Author(s)

Mikis Stasinopoulos, Bob Rigby, Marco Enea and Fernanda de Bastiani

References

Efron, B., 1986. Double exponential families and their use in generalized linear Regression. Journal of the American Statistical Association 81 (395), 709-721.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

BI,BB

Examples

DBI()
x <- 0:20
# underdispersed DBI
plot(x, dDBI(x, mu=.5, sigma=.2, bd=20), type="h", col="green", lwd=2)
# binomial
lines(x+0.1, dDBI(x, mu=.5, sigma=1, bd=20), type="h", col="black", lwd=2)
# overdispersed DBI
lines(x+.2, dDBI(x, mu=.5, sigma=2, bd=20), type="h", col="red",lwd=2)

[Package gamlss.dist version 5.1-5 Index]