DBURR12 {gamlss.dist}	R Documentation

The Discrete Burr type XII distribution for fitting a GAMLSS model

Description

The DBURR12() function defines the discrete Burr type XII distribution, a three parameter discrete distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dDBURR12(), pDBURR12(), qDBURR12() and rDBURR12() define the density, distribution function, quantile function and random generation for the discrete Burr type XII DBURR12(), distribution.

Usage

DBURR12(mu.link = "log", sigma.link = "log", nu.link = "log")
dDBURR12(x, mu = 5, sigma = 2, nu = 2, log = FALSE)
pDBURR12(q, mu = 5, sigma = 2, nu = 2, lower.tail = TRUE, 
        log.p = FALSE)
qDBURR12(p, mu = 5, sigma = 2, nu = 2, lower.tail = TRUE, 
        log.p = FALSE)
rDBURR12(n, mu = 5, sigma = 2, nu = 2)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter

nu.link
Defines the nu.link, with "log" link as the default for the nu parameter

x
vector of (non-negative integer) quantiles

p
vector of probabilities

q
vector of quantiles

mu
vector of positive mu

sigma
vector of positive dispersion parameter sigma

nu
vector of nu

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

n
number of random values to return

Details

The probability function of the discrete Burr XII distribution is given by

f(y|μ,σ,ν)= (1+(y/μ)^σ)^ν - (1+((y+1)/μ)^σ)^ν

for y=0,1,2,..., mu>0 , σ>0 and nu>0.

Note that the above parametrization is different from Para and Jan (2016).

Value

The function DBURR12() Returns a gamlss.family object which can be used to fit a discrete Burr XII distribution in the gamlss() function.

Note

The parameters of the distributioins are highly correlated so the argument of gamlss method=mixed(10,100) may have to be used.

The distribution can be under/over dispersed and also with long tails.

Author(s)

Rigby, R. A., Stasinopoulos D. M., Fernanda De Bastiani.

References

Para, B. A. and Jan, T. R. (2016). On discrete three parameter Burr type XII and discrete Lomax distributions and their applications to model count data from medical science. Biometrics and Biostatistics International Journal, 54, part 3, pp 507-554.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 4, pp 1-15.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, DPO

Examples

DBURR12()# 
#plot the pdf using plot 
plot(function(y) dDBURR12(y, mu=10, sigma=1, nu=1), from=0, to=100, n=100+1, type="h") # pdf
# plot the cdf
plot(seq(from=0,to=100),pDBURR12(seq(from=0,to=100), mu=10, sigma=1, nu=1), type="h")   # cdf
# generate random sample
tN <- table(Ni <- rDBURR12(100, mu=5, sigma=1, nu=1))
r <- barplot(tN, col='lightblue')

[Package gamlss.dist version 5.1-5 Index]