EGB2 {gamlss.dist}	R Documentation

The exponential generalized Beta type 2 distribution for fitting a GAMLSS

Description

This function defines the generalized t distribution, a four parameter distribution. The response variable is in the range from minus infinity to plus infinity. The functions dEGB2, pEGB2, qEGB2 and rEGB2 define the density, distribution function, quantile function and random generation for the generalized beta type 2 distribution.

Usage

EGB2(mu.link = "identity", sigma.link = "log", nu.link = "log", 
      tau.link = "log")
dEGB2(x, mu = 0, sigma = 1, nu = 1, tau = 0.5, log = FALSE)
pEGB2(q, mu = 0, sigma = 1, nu = 1, tau = 0.5, lower.tail = TRUE, 
      log.p = FALSE)
qEGB2(p, mu = 0, sigma = 1, nu = 1, tau = 0.5, lower.tail = TRUE, 
      log.p = FALSE)
rEGB2(n, mu = 0, sigma = 1, nu = 1, tau = 0.5)

Arguments

mu.link
Defines the mu.link, with "identity" link as the default for the mu parameter.

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter.

nu.link
Defines the nu.link, with "log" link as the default for the nu parameter.

tau.link
Defines the tau.link, with "log" link as the default for the tau parameter.

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

nu
vector of skewness nu parameter values

tau
vector of kurtosis tau parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The probability density function of the Generalized Beta type 2, (GB2), is defined as

f(y|mu,sigma,nu,tau)=exp{nu*z}(abs(sigma)*Beta(nu.tau)*(1+exp(z))^(nu+tau) )^(-1)

for -Inf<y<Inf, where z=(y-mu)/sigma and -Inf<mu<Inf, -Inf<sigma<Inf, nu>0 and tau>0, McDonald and Xu (1995).

Value

EGB2() returns a gamlss.family object which can be used to fit the EGB2 distribution in the gamlss() function. dEGB2() gives the density, pEGB2() gives the distribution function, qEGB2() gives the quantile function, and rEGB2() generates random deviates.

Author(s)

Bob Rigby and Mikis Stasinopoulos mikis.stasinopoulos@gamlss.org

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R. A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, JSU, BCT

Examples

EGB2()   # 
y<- rEGB2(200, mu=5, sigma=2, nu=1, tau=4)
library(MASS)
truehist(y)
fx<-dEGB2(seq(min(y), 20, length=200), mu=5 ,sigma=2, nu=1, tau=4)
lines(seq(min(y),20,length=200),fx)
# something funny here
# library(gamlss)
# histDist(y, family=EGB2, n.cyc=60)
integrate(function(x) x*dEGB2(x=x, mu=5, sigma=2, nu=1, tau=4), -Inf, Inf)
curve(dEGB2(x, mu=5 ,sigma=2, nu=1, tau=4), -10, 10, main = "The EGB2  density 
             mu=5, sigma=2, nu=1, tau=4")

[Package gamlss.dist version 5.1-4 Index]