GA {gamlss.dist}	R Documentation

Gamma distribution for fitting a GAMLSS

Description

The function GA defines the gamma distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The parameterization used has the mean of the distribution equal to mu and the variance equal to (sigma^2)*(mu^2). The functions dGA, pGA, qGA and rGA define the density, distribution function, quantile function and random generation for the specific parameterization of the gamma distribution defined by function GA.

Usage

GA(mu.link = "log", sigma.link ="log")
dGA(x, mu = 1, sigma = 1, log = FALSE) 
pGA(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qGA(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rGA(n, mu = 1, sigma = 1)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse", "identity" ans "own"

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter, other link is the "inverse", "identity" and "own"

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The specific parameterization of the gamma distribution used in GA is

f(y|mu,sigma)=(y^((1/sigma^2)-1)*exp[-y/((sigma^2)*mu)])/((sigma^2*mu)^(1/sigma^2) Gamma(1/sigma^2))

for y>0, μ>0 and σ>0.

Value

GA() returns a gamlss.family object which can be used to fit a gamma distribution in the gamlss() function. dGA() gives the density, pGA() gives the distribution function, qGA() gives the quantile function, and rGA() generates random deviates. The latest functions are based on the equivalent R functions for gamma distribution.

Note

mu is the mean of the distribution in GA. In the function GA, sigma is the square root of the usual dispersion parameter for a GLM gamma model. Hence sigma*mu is the standard deviation of the distribution defined in GA.

Author(s)

Mikis Stasinopoulos mikis.stasinopoulos@gamlss.org, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family

Examples

GA()# gives information about the default links for the gamma distribution      
# dat<-rgamma(100, shape=1, scale=10) # generates 100 random observations 
# fit a gamlss model
# gamlss(dat~1,family=GA) 
# fits a constant for each parameter mu and sigma of the gamma distribution
newdata<-rGA(1000,mu=1,sigma=1) # generates 1000 random observations
hist(newdata) 
rm(dat,newdata)

[Package gamlss.dist version 5.1-4 Index]