GAF {gamlss.dist}	R Documentation

The Gamma distribution family

Description

The function GAF() defines a gamma distribution family, which has three parameters. This is not the generalised gamma distribution which is called GG. The third parameter here is to model the mean and variance relationship. The distribution can be fitted using the function gamlss(). The mean of GAF is equal to mu. The variance is equal to sigma^2*mu^nu so the standard deviation is sigma*mu^(nu/2). The function is design for cases where the variance is proportional to a power of the mean. This is an instance of the Taylor's power low, see Enki et al. (2017). The functions dGAF, pGAF, qGAF and rGAF define the density, distribution function, quantile function and random generation for the GAF parametrization of the gamma family.

Usage

GAF(mu.link = "log", sigma.link = "log", nu.link = "identity")
dGAF(x, mu = 1, sigma = 1, nu = 2, log = FALSE)
pGAF(q, mu = 1, sigma = 1, nu = 2, lower.tail = TRUE, 
    log.p = FALSE)
qGAF(p, mu = 1, sigma = 1, nu = 2, lower.tail = TRUE, 
    log.p = FALSE)
rGAF(n, mu = 1, sigma = 1, nu = 2)

Arguments

mu.link
Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter

nu.link
Defines the nu.link with "identity" link as the default for the nu parameter

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

nu
vector of power parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parametrization of the gamma family given in the function GAF() is:

f(y|μ,σ_1)=(y^((1/σ_1^2)-1)*exp[-y/((σ_1^2)*μ)])/((σ_1^2*μ)^(1/σ_1^2) Gamma(1/σ_1^2))

for y>0, μ>0 where σ_1=σ μ^(ν/1-1) .

Value

GAF() returns a gamlss.family object which can be used to fit the gamma family in the gamlss() function.

Note

For the function GAF(), μ is the mean and σ μ^{ν/2} is the standard deviation of the gamma family. The GAF is design for fitting regression type models where the variance is proportional to a power of the mean.

Note that because the high correlation between the sigma and the nu parameter the mixed() method should be used in the fitting.

Author(s)

Mikis Stasinopoulos, Robert Rigby and Fernanda De Bastiani

References

Enki, D G, Noufaily, A., Farrington, P., Garthwaite, P., Andrews, N. and Charlett, A. (2017) Taylor's power law and the statistical modelling of infectious disease surveillance data, Journal of the Royal Statistical Society: Series A (Statistics in Society), volume=180, number=1, pages=45-72.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, GA, GG

Examples

GAF()
## Not run: 
m1<-gamlss(y~poly(x,2),data=abdom,family=GAF, method=mixed(1,100),  
           c.crit=0.00001)
# using RS()
m2<-gamlss(y~poly(x,2),data=abdom,family=GAF,  n.cyc=5000, c.crit=0.00001)
# the estimates of nu slightly different
fitted(m1, "nu")[1]
fitted(m2, "nu")[1]
# global deviance almost identical
AIC(m1, m2)

## End(Not run)

[Package gamlss.dist version 5.1-4 Index]