GB1 {gamlss.dist}	R Documentation

The generalized Beta type 1 distribution for fitting a GAMLSS

Description

This function defines the generalized beta type 1 distribution, a four parameter distribution. The function GB1 creates a gamlss.family object which can be used to fit the distribution using the function gamlss(). Note the range of the response variable is from zero to one. The functions dGB1, GB1, qGB1 and rGB1 define the density, distribution function, quantile function and random generation for the generalized beta type 1 distribution.

Usage

GB1(mu.link = "logit", sigma.link = "logit", nu.link = "log", 
      tau.link = "log")
dGB1(x, mu = 0.5, sigma = 0.4, nu = 1, tau = 1, log = FALSE)
pGB1(q, mu = 0.5, sigma = 0.4, nu = 1, tau = 1, lower.tail = TRUE, 
       log.p = FALSE)
qGB1(p, mu = 0.5, sigma = 0.4, nu = 1, tau = 1, lower.tail = TRUE, 
       log.p = FALSE)
rGB1(n, mu = 0.5, sigma = 0.4, nu = 1, tau = 1)

Arguments

mu.link
Defines the mu.link, with "identity" link as the default for the mu parameter.

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter.

nu.link
Defines the nu.link, with "log" link as the default for the nu parameter.

tau.link
Defines the tau.link, with "log" link as the default for the tau parameter.

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

nu
vector of skewness nu parameter values

tau
vector of kurtosis tau parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The probability density function of the Generalized Beta type 1, (GB1), is defined as

f(y|mu,sigma,nu,tau)=(tau*nu^beta*y^(tau*alpha-1)(1-y^tau))^(beta-1)/(Beta(alpha,beta)*(nu+(1-nu)*y^tau))^(alpha*beta))

where 0<y<1, alpha=mu*(1-sigma^2)/sigma^2 and (1-mu)*(1-sigma^2)/sigma^2, and alpha>0, beta>0. Note the alpha/(alpha+beta), sigma=(alpha+beta+1)^(-1/2). .

Value

GB1() returns a gamlss.family object which can be used to fit the GB1 distribution in the gamlss() function. dGB1() gives the density, pGB1() gives the distribution function, qGB1() gives the quantile function, and rGB1() generates random deviates.

Warning

The qSHASH and rSHASH are slow since they are relying on golden section for finding the quantiles

Author(s)

Bob Rigby and Mikis Stasinopoulos mikis.stasinopoulos@gamlss.org

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R. A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, JSU, BCT

Examples

GB1()   # 
y<- rGB1(200, mu=.1, sigma=.6, nu=1, tau=4)
hist(y)
# library(gamlss)
# histDist(y, family=GB1, n.cyc=60)
curve(dGB1(x, mu=.1 ,sigma=.6, nu=1, tau=4), 0.01, 0.99, main = "The GB1  
           density mu=0.1, sigma=.6, nu=1, tau=4")

[Package gamlss.dist version 5.0-5 Index]