IG {gamlss.dist}	R Documentation

Inverse Gaussian distribution for fitting a GAMLSS

Description

The function IG(), or equivalently Inverse.Gaussian(), defines the inverse Gaussian distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dIG, pIG, qIG and rIG define the density, distribution function, quantile function and random generation for the specific parameterization of the Inverse Gaussian distribution defined by function IG.

Usage

IG(mu.link = "log", sigma.link = "log")
dIG(x, mu = 1, sigma = 1, log = FALSE)
pIG(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qIG(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rIG(n, mu = 1, sigma = 1, ...)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

...
... can be used to pass the uppr.limit argument to qIG

Details

Definition file for inverse Gaussian distribution.

f(y|mu,sigma)=(1/(sqrt(2*pi*sigma^2*y^3))* exp(-(y-mu)^2/(2*mu^2*sigma^2*y))

for y>0, μ>0 and σ>0.

Value

returns a gamlss.family object which can be used to fit a inverse Gaussian distribution in the gamlss() function.

Note

mu is the mean and (sigma^2)*(mu^3) is the variance of the inverse Gaussian

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/)

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family , GA, GIG

Examples

IG()# gives information about the default links for the normal distribution
# library(gamlss)
# data(rent)        
# gamlss(R~cs(Fl),family=IG, data=rent) # 
plot(function(x)dIG(x, mu=1,sigma=.5), 0.01, 6, 
 main = "{Inverse Gaussian  density mu=1,sigma=0.5}")
plot(function(x)pIG(x, mu=1,sigma=.5), 0.01, 6, 
 main = "{Inverse Gaussian  cdf mu=1,sigma=0.5}")

[Package gamlss.dist version 5.1-4 Index]