IGAMMA {gamlss.dist}	R Documentation

Inverse Gamma distribution for fitting a GAMLSS

Description

The function IGAMMA() defines the Inverse Gamma distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(), with parameters mu (the mode) and sigma. The functions dIGAMMA, pIGAMMA, qIGAMMA and rIGAMMA define the density, distribution function, quantile function and random generation for the IGAMMA parameterization of the Inverse Gamma distribution.

Usage

IGAMMA(mu.link = "log", sigma.link="log")
dIGAMMA(x, mu = 1, sigma = .5, log = FALSE)
pIGAMMA(q, mu = 1, sigma = .5, lower.tail = TRUE, log.p = FALSE)
qIGAMMA(p, mu = 1, sigma = .5, lower.tail = TRUE, log.p = FALSE)
rIGAMMA(n, mu = 1, sigma = .5)

Arguments

mu.link
Defines the mu.link, with log link as the default for the mu parameter

sigma.link
Defines the sigma.link, with log as the default for the sigma parameter

x, q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise P[X > x]

p
vector of probabilities

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the Inverse Gamma distribution in the function IGAMMA is

f(y|mu, sigma) = ([mu (alpha+1)]^alpha)/Gamma(alpha) y^(-(alpha+1)) exp(-(mu (alpha+1))/y)

where alpha = 1/(sigma^2) for y>0, mu>0 and sigma>0.

Value

returns a gamlss.family object which can be used to fit an Inverse Gamma distribution in the gamlss() function.

Note

For the function IGAMMA(), mu is the mode of the Inverse Gamma distribution.

Author(s)

Fiona McElduff, Bob Rigby and Mikis Stasinopoulos.

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, GA

Examples

par(mfrow=c(2,2))
y<-seq(0.2,20,0.2)
plot(y, dIGAMMA(y), type="l")
q <- seq(0.2, 20, 0.2)
plot(q, pIGAMMA(q), type="l")
p<-seq(0.0001,0.999,0.05)
plot(p , qIGAMMA(p), type="l")
dat <- rIGAMMA(50)
hist(dat)
#summary(gamlss(dat~1, family="IGAMMA"))

[Package gamlss.dist version 5.1-4 Index]