LG {gamlss.dist}	R Documentation

Logarithmic and zero adjusted logarithmic distributions for fitting a GAMLSS model

Description

The function LG defines the logarithmic distribution, a one parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dLG, pLG, qLG and rLG define the density, distribution function, quantile function and random generation for the logarithmic , LG(), distribution.

The function ZALG defines the zero adjusted logarithmic distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dZALG, pZALG, qZALG and rZALG define the density, distribution function, quantile function and random generation for the inflated logarithmic , ZALG(), distribution.

Usage

LG(mu.link = "logit")
dLG(x, mu = 0.5, log = FALSE)
pLG(q, mu = 0.5, lower.tail = TRUE, log.p = FALSE)
qLG(p, mu = 0.5, lower.tail = TRUE, log.p = FALSE, max.value = 10000)
rLG(n, mu = 0.5)
ZALG(mu.link = "logit", sigma.link = "logit")
dZALG(x, mu = 0.5, sigma = 0.1, log = FALSE)
pZALG(q, mu = 0.5, sigma = 0.1, lower.tail = TRUE, log.p = FALSE)
qZALG(p, mu = 0.5, sigma = 0.1, lower.tail = TRUE, log.p = FALSE)
rZALG(n, mu = 0.5, sigma = 0.1)

Arguments

mu.link
defines the mu.link, with logit link as the default for the mu parameter

sigma.link
defines the sigma.link, with logit link as the default for the sigma parameter which in this case is the probability at zero.

x
vector of (non-negative integer)

mu
vector of positive means

sigma
vector of probabilities at zero

p
vector of probabilities

q
vector of quantiles

n
number of random values to return

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

max.value
valued needed for the numerical calculation of the q-function

Details

For the definition of the distributions see Rigby and Stasinopoulos (2010) below.

The parameterization of the logarithmic distribution in the function LM is

f(y|mu) = α μ^y / y

where for y>=1 and μ>0 and

α= [log(1-μ)]^{-1}

Value

The function LG and ZALG return a gamlss.family object which can be used to fit a logarithmic and a zero inflated logarithmic distributions respectively in the gamlss() function.

Author(s)

Mikis Stasinopoulos, Bob Rigby

References

Johnson, Norman Lloyd; Kemp, Adrienne W; Kotz, Samuel (2005). "Chapter 7: Logarithmic and Lagrangian distributions". Univariate discrete distributions (3 ed.). John Wiley & Sons. ISBN 9780471272465.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) Distributions for modeling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC. An older version can be found in http://www.gamlss.com/.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Rigby, R. A. and Stasinopoulos D. M. (2010) The gamlss.family distributions, (distributed with this package or see http://www.gamlss.com/)

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, PO, ZAP

Examples

LG() 
ZAP()
# creating data and plotting them 
 dat <- rLG(1000, mu=.3)
   r <- barplot(table(dat), col='lightblue')
dat1 <- rZALG(1000, mu=.3, sigma=.1)
  r1 <- barplot(table(dat1), col='lightblue')

[Package gamlss.dist version 5.1-5 Index]