LO(gamlss.dist)	R Documentation

Logistic distribution for fitting a GAMLSS

Description

The function LO(), or equivalently Logistic(), defines the logistic distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss()

Usage


LO(mu.link = "identity", sigma.link = "log")
dLO(x, mu = 0, sigma = 1, log = FALSE)
pLO(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qLO(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rLO(n, mu = 0, sigma = 1)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

Definition file for Logistic distribution.

f(y|mu,sigma)=(1/sigma)*exp((y-mu)/sigma)*((1+exp(-(y-mu)/sigma))^(-2))

for y=(-Inf,+Inf), μ=(-Inf,+Inf) and σ>0.

Value

LO() returns a gamlss.family object which can be used to fit a logistic distribution in the gamlss() function. dLO() gives the density, pLO() gives the distribution function, qLO() gives the quantile function, and rLO() generates random deviates for the logistic distribution. The latest functions are based on the equivalent R functions for logistic distribution.

Note

mu is the mean and sigma*pi/sqrt(3) is the standard deviation for the logistic distribution

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, NO, TF

Examples


LO()# gives information about the default links for the Logistic distribution 
plot(function(y) dLO(y, mu=10 ,sigma=2), 0, 20)
plot(function(y) pLO(y, mu=10 ,sigma=2), 0, 20)
plot(function(y) qLO(y, mu=10 ,sigma=2), 0, 1)
# library(gamlss)
# data(abdom)
# h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=LO, data=abdom) # fits 
# plot(h)

[Package gamlss.dist version 3.0-0 Index]