LQNO {gamlss.dist}	R Documentation

Normal distribution with a specific mean and variance relationship for fitting a GAMLSS model

Description

The function LQNO() defines a normal distribution family, which has a specific mean and variance relationship. The distribution can be used in a GAMLSS fitting using the function gamlss(). The mean of LQNO is equal to mu. The variance is equal to mu*(1+sigma*mu) so the standard deviation is sqrt(mu*(1+sigma*mu)). The function is found useful in modelling small RNA sequencing experiments. The functions dLQNO, pLQNO, qLQNO and rLQNO define the density, distribution function, quantile function (inverse cdf) and random generation for the LQNO() parametrization of the normal distribution.

Usage

LQNO(mu.link = "log", sigma.link = "log")
dLQNO(x, mu = 1, sigma = 1, log = FALSE)
pLQNO(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qLQNO(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rLQNO(n, mu = 1, sigma = 1)

Arguments

mu.link
mu link function with "log" as default

sigma.link
mu link function with "log" as default

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

p
vector of probabilities

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

LQNO stands for Linear Quadratic Normal Family, in which the variance is a linear quadratic function of the mean: Var(Y) = mu*(1+sigma*mu). This is created to facilitate the analysis of data coming from small RNA sequencing experiments, basically counts of short RNAs that one isolates from cells or biofluids such as urine, plasma or cerebrospinal fluid. Argyropoulos et al. (2017) showing that the LQNO distribution (and the Negative Binomial which implements the same mean- variance relationship) are highly accurate approximations to the generative models of the signals in these experiments

Value

The function LQNO returns a gamlss.family object which can be used to fit this specific form of the normal distribution family in the gamlss() function.

Note

The mu parameters must be positive so for the relationship Var(Y) = mu*(1+sigma*mu) to be valid.

Author(s)

Christos Argyropoulos

References

Argyropoulos C, Etheridge A, Sakhanenko N, Galas D. (2017) Modeling bias and variation in the stochastic processes of small RNA sequencing. Nucleic Acids Res. 2017 Mar 27. doi: 10.1093/nar/gkx199. [Epub ahead of print] PubMed PMID: 28369495.

See Also

NO,NO2, NOF

Examples

LQNO()# gives information about the default links for the normal distribution
# a comparison of different Normal models
#m1 <- gamlss(y~pb(x), sigma.fo=~pb(x), data=abdom,  family=NO(mu.link="log"))
#m2 <- gamlss(y~pb(x), sigma.fo=~pb(x), data=abdom, family=LQNO)
#m3 <- gamlss(y~pb(x), sigma.fo=~pb(x), data=abdom, family=NOF(mu.link="log"))
#AIC(m1,m2,m3)

[Package gamlss.dist version 5.0-5 Index]