NO2 {gamlss.dist}	R Documentation

Normal distribution (with variance as sigma parameter) for fitting a GAMLSS

Description

The function NO2() defines the normal distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss() with mean equal to mu and variance equal to sigma. The functions dNO2, pNO2, qNO2 and rNO2 define the density, distribution function, quantile function and random generation for this specific parameterization of the normal distribution.

[A alternative parameterization with sigma as the standard deviation is given in the function NO()]

Usage

NO2(mu.link = "identity", sigma.link = "log")
dNO2(x, mu = 0, sigma = 1, log = FALSE)
pNO2(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qNO2(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rNO2(n, mu = 0, sigma = 1)

Arguments

mu.link
Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parametrization of the normal distribution given in the function NO2() is

f(y|mu,sigma)=(1/(sqrt(2*pi*sigma))* exp(-0.5*(y-mu)^2/sigma)

for y=(-Inf,+Inf), μ=(-Inf,+Inf) and σ>0.

Value

returns a gamlss.family object which can be used to fit a normal distribution in the gamlss() function.

Note

For the function NO(), mu is the mean and sigma is the standard deviation (not the variance) of the normal distribution. [The function NO2() defines the normal distribution with sigma as the variance.]

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, NO

Examples

NO()# gives information about the default links for the normal distribution
dat<-rNO(100)
hist(dat)        
plot(function(y) dNO(y, mu=10 ,sigma=2), 0, 20)
plot(function(y) pNO(y, mu=10 ,sigma=2), 0, 20)
plot(function(y) qNO(y, mu=10 ,sigma=2), 0, 1)
# library(gamlss)
# gamlss(dat~1,family=NO) # fits a constant for mu and sigma 

[Package gamlss.dist version 5.0-5 Index]