PO {gamlss.dist}	R Documentation

Poisson distribution for fitting a GAMLSS model

Description

This function PO defines the Poisson distribution, an one parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dPO, pPO, qPO and rPO define the density, distribution function, quantile function and random generation for the Poisson, PO(), distribution.

Usage

PO(mu.link = "log")
dPO(x, mu = 1, log = FALSE)
pPO(q, mu = 1, lower.tail = TRUE, log.p = FALSE)
qPO(p, mu = 1, lower.tail = TRUE, log.p = FALSE)
rPO(n, mu = 1)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter

x
vector of (non-negative integer) quantiles

mu
vector of positive means

p
vector of probabilities

q
vector of quantiles

n
number of random values to return

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

Details

Definition file for Poisson distribution.

f(y|μ)=e^(-μ)*μ^y/Γ(y+1)

for y=0,1,2,... and μ>0.

Value

returns a gamlss.family object which can be used to fit a Poisson distribution in the gamlss() function.

Note

mu is the mean of the Poisson distribution

Author(s)

Bob Rigby, Mikis Stasinopoulos mikis.stasinopoulos@gamlss.org, and Kalliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, NBI, NBII, SI, SICHEL

Examples

PO()# gives information about the default links for the Poisson distribution  
# fitting data using PO()

# plotting the distribution
plot(function(y) dPO(y, mu=10 ), from=0, to=20, n=20+1, type="h")
# creating random variables and plot them 
tN <- table(Ni <- rPO(1000, mu=5))
 r <- barplot(tN, col='lightblue')
# library(gamlss)
# data(aids)
# h<-gamlss(y~cs(x,df=7)+qrt, family=PO, data=aids) # fits the constant+x+qrt model 
# plot(h)
# pdf.plot(family=PO, mu=10, min=0, max=20, step=1)

[Package gamlss.dist version 5.1-4 Index]