RG(gamlss.dist)	R Documentation

The Reverse Gumbel distribution for fitting a GAMLSS

Description

The function RG defines the reverse Gumbel distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dRG, pRG, qRG and rRG define the density, distribution function, quantile function and random generation for the specific parameterization of the reverse Gumbel distribution.

Usage


RG(mu.link = "identity", sigma.link = "log")
dRG(x, mu = 0, sigma = 1, log = FALSE)
pRG(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qRG(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rRG(n, mu = 0, sigma = 1)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter. other available link is "inverse", "log" and "own"

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter, other links are the "inverse", "identity" and "own"

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

The specific parameterization of the reverse Gumbel distribution used in RG is

f(y|mu,sigma)=(1/sigma)*exp(-((y-mu)/sigma)-exp(-((y-mu)/sigma)))

for y=(-Inf,+Inf), μ=(-Inf,+Inf) and σ>0.

Value

RG() returns a gamlss.family object which can be used to fit a Gumbel distribution in the gamlss() function. dRG() gives the density, pGU() gives the distribution function, qRG() gives the quantile function, and rRG() generates random deviates.

Note

The mean of the distribution is mu-0.57722*sigma and the variance is pi^2 sigma^2/6.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

Examples



plot(function(x) dRG(x, mu=0,sigma=1), -3, 6, 
 main = "{Reverse Gumbel  density mu=0,sigma=1}")
RG()# gives information about the default links for the Gumbel distribution      
dat<-rRG(100, mu=10, sigma=2) # generates 100 random observations 
# library(gamlss)
# gamlss(dat~1,family=RG) # fits a constant for each parameter mu and sigma 


[Package gamlss.dist version 3.0-0 Index]