SIMPLEX {gamlss.dist}	R Documentation

The simplex distribution for fitting a GAMLSS

Description

The functions SIMPLEX() define the simplex distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). SIMPLEX() has mean equal to the parameter mu and sigma as scale parameter, see below. The functions dSIMPLEX, pSIMPLEX qSIMPLEX and rSIMPLEX define the density, comulative distribution function, quantile function and random generation for the simplex distribution.

Usage

SIMPLEX(mu.link = "logit", sigma.link = "log")
dSIMPLEX(x, mu = 0.5, sigma = 1, log = FALSE)
pSIMPLEX(q, mu = 0.5, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qSIMPLEX(p, mu = 0.5, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rSIMPLEX(n = 1, mu = 0.5, sigma = 1)

Arguments

mu.link
the mu link function with default logit

sigma.link
the sigma link function with default log

x,q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The simplex distribution is given as

f(y|mu,sigma) = 1/(2*pi*sigma^2*(y*(1-y))^3)^(1/2) exp(-(1/(2*sigma^2)) * ((y-mu)^2)/(y*(1-y)*mu^2*(1-mu)^2))

for y in (0,1), 0<μ<1 and σ>0.

Value

SIMPLEX() returns a gamlss.family object which can be used to fit a simplex distribution in the gamlss() function.

Author(s)

Bob Rigby, Mikis Stasinopoulos and Fernanda De Bastiani

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

Examples

SIMPLEX()#  default links for the simplex distribution
plot(function(y) dSIMPLEX(y, mu=.5 ,sigma=1), 0.001, .999)
plot(function(y) pSIMPLEX(y, mu=.5 ,sigma=1), 0.001, 0.999)
plot(function(y) qSIMPLEX(y, mu=.5 ,sigma=1), 0.001, 0.999)
plot(function(y) qSIMPLEX(y, mu=.5 ,sigma=1, lower.tail=FALSE), 0.001, .999)

[Package gamlss.dist version 5.1-4 Index]