SN2 {gamlss.dist}	R Documentation

Skew Normal Type 2 distribution for fitting a GAMLSS

Description

The function SN2() defines the Skew Normal Type 2 distribution, a three parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(), with parameters mu, sigma and nu. The functions dSN2, pSN2, qSN2 and rSN2 define the density, distribution function, quantile function and random generation for the SN2 parameterization of the Skew Normal Type 2 distribution.

Usage

SN2(mu.link = "identity", sigma.link = "log", nu.link = "log")
dSN2(x, mu = 0, sigma = 1, nu = 2, log = FALSE)
pSN2(q, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)
qSN2(p, mu = 0, sigma = 1, nu = 2, lower.tail = TRUE, log.p = FALSE)
rSN2(n, mu = 0, sigma = 1, nu = 2)

Arguments

mu.link
Defines the mu.link, with "‘identity"’ links the default for the mu parameter

sigma.link
Defines the sigma.link, with "‘log"’ as the default for the sigma parameter

nu.link
Defines the nu.link, with "‘log"’ as the default for the sigma parameter

x, q
vector of quantiles

mu
vector of location parameter values

sigma
vector of scale parameter values

nu
vector of scale parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise P[X > x]

p
vector of probabilities

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the Skew Normal Type 2 distribution in the function SN2 is ...

Value

returns a gamlss.family object which can be used to fit a Skew Normal Type 2 distribution in the gamlss() function.

Note

This is a special case of the Skew Exponential Power type 3 distribution (SEP3)where tau=2.

Author(s)

Mikis Stasinopoulos, Bob Rigby and Fiona McElduff.

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) Distributions for modeling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC. An older version can be found in http://www.gamlss.com/.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family

Examples

par(mfrow=c(2,2))
y<-seq(-3,3,0.2)
plot(y, dSN2(y), type="l" , lwd=2)
q<-seq(-3,3,0.2)
plot(q, pSN2(q), ylim=c(0,1), type="l", lwd=2) 
p<-seq(0.0001,0.999,0.05)
plot(p, qSN2(p), type="l", lwd=2)
dat <- rSN2(100)
hist(rSN2(100), nclass=30)

[Package gamlss.dist version 5.1-5 Index]