WEI {gamlss.dist}	R Documentation

Weibull distribution for fitting a GAMLSS

Description

The function WEI can be used to define the Weibull distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). [Note that the GAMLSS function WEI2 uses a different parameterization for fitting the Weibull distribution.] The functions dWEI, pWEI, qWEI and rWEI define the density, distribution function, quantile function and random generation for the specific parameterization of the Weibul distribution.

Usage

WEI(mu.link = "log", sigma.link = "log")
dWEI(x, mu = 1, sigma = 1, log = FALSE)
pWEI(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qWEI(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rWEI(n, mu = 1, sigma = 1)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse", "identity" and "own"

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter, other link is the "inverse", "identity" and "own"

x,q
vector of quantiles

mu
vector of the mu parameter

sigma
vector of sigma parameter

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the function WEI is given by

f(y|mu,sigma)= (sigma*y^(sigma-1)/mu^sigma)exp(-(y/mu)^sigma)

for y>0, μ>0 and σ>0. The GAMLSS functions dWEI, pWEI, qWEI, and rWEI can be used to provide the pdf, the cdf, the quantiles and random generated numbers for the Weibull distribution with argument mu, and sigma. [See the GAMLSS function WEI2 for a different parameterization of the Weibull.]

Value

WEI() returns a gamlss.family object which can be used to fit a Weibull distribution in the gamlss() function. dWEI() gives the density, pWEI() gives the distribution function, qWEI() gives the quantile function, and rWEI() generates random deviates. The latest functions are based on the equivalent R functions for Weibull distribution.

Note

The mean in WEI is given by mu*gamma((1/sigma)+1) and the variance (mu^2)*(gamma((2/sigma)+1)-gamma((1/sigma)+1)^2)

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) Distributions for modeling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC. An older version can be found in http://www.gamlss.com/.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.com/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, WEI2, WEI3

Examples

WEI()
dat<-rWEI(100, mu=10, sigma=2)
# library(gamlss)
# gamlss(dat~1, family=WEI)

[Package gamlss.dist version 5.1-5 Index]