WEI3 {gamlss.dist}	R Documentation

A specific parameterization of the Weibull distribution for fitting a GAMLSS

Description

The function WEI3 can be used to define the Weibull distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). This is a parameterization of the Weibull distribution where mu is the mean of the distribution. [Note that the GAMLSS functions WEI and WEI2 use different parameterizations for fitting the Weibull distribution.] The functions dWEI3, pWEI3, qWEI3 and rWEI3 define the density, distribution function, quantile function and random generation for the specific parameterization of the Weibull distribution.

Usage

WEI3(mu.link = "log", sigma.link = "log")
dWEI3(x, mu = 1, sigma = 1, log = FALSE)
pWEI3(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qWEI3(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rWEI3(n, mu = 1, sigma = 1)

Arguments

mu.link
Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse" and "identity"

sigma.link
Defines the sigma.link, with "log" link as the default for the sigma parameter, other link is the "inverse" and "identity"

x,q
vector of quantiles

mu
vector of the mu parameter values

sigma
vector of sigma parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the function WEI3 is given by

f(y|mu,sigma)= (sigma/beta)*(y/beta)^(sigma-1)exp(-(y/beta)^sigma)

where mu/(Gamma((1/sigma)+1)) for y>0, μ>0 and σ>0. The GAMLSS functions dWEI3, pWEI3, qWEI3, and rWEI3 can be used to provide the pdf, the cdf, the quantiles and random generated numbers for the Weibull distribution with argument mu, and sigma. [See the GAMLSS function WEI for a different parameterization of the Weibull.]

Value

WEI3() returns a gamlss.family object which can be used to fit a Weibull distribution in the gamlss() function. dWEI3() gives the density, pWEI3() gives the distribution function, qWEI3() gives the quantile function, and rWEI3() generates random deviates. The latest functions are based on the equivalent R functions for Weibull distribution.

Warning

In WEI3 the estimated parameters mu and sigma can be highly correlated so it is advisable to use the CG() method for fitting [as the RS() method can be very slow in this situation.]

Note

The mean in WEI3 is given by mu and the variance mu^2(Gamma(1/sigma +1)/Gamma(1/sigma+1)^2 -1)

Author(s)

Bob Rigby and Mikis Stasinopoulos

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) Distributions for modeling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC. An older version can be found in http://www.gamlss.com/.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family, WEI, WEI2

Examples

WEI3()
dat<-rWEI(100, mu=.1, sigma=2)
# library(gamlss)
# gamlss(dat~1, family=WEI3, method=CG())

[Package gamlss.dist version 5.1-5 Index]