gamlss.dist-package {gamlss.dist}	R Documentation

Distributions for Generalized Additive Models for Location Scale and Shape

Description

A set of distributions which can be used for modelling the response variables in Generalized Additive Models for Location Scale and Shape, Rigby and Stasinopoulos (2005), <doi:10.1111/j.1467-9876.2005.00510.x>. The distributions can be continuous, discrete or mixed distributions. Extra distributions can be created, by transforming, any continuous distribution defined on the real line, to a distribution defined on ranges 0 to infinity or 0 to 1, by using a ”log” or a ”logit' transformation respectively.

Details

The DESCRIPTION file:

Package: 	 gamlss.dist

Type: 	 Package

Title: 	 Distributions for Generalized Additive Models for Location Scale and Shape

Version: 	 5.1-4

Date: 	 2019-5-14

Authors@R: 	 c(person("Mikis", "Stasinopoulos", role = c("aut", "cre", "cph"), email = "d.stasinopoulos@londonmet.ac.uk"), person("Robert", "Rigby", role = c("aut"), email = "r.rigby@londonmet.ac.uk"), person("Calliope", "Akantziliotou", role = "ctb"), person("Vlasios", "Voudouris", role = "ctb"), person("Gillian", "Heller", role = "ctb"), person("Fernanda", "De Bastiani", role = "ctb"), person("Raydonal", "Ospina", role = "ctb", email= "rospina@ime.usp.br"), person("Nicoletta", "Motpan", role = "ctb"), person("Fiona", "McElduff", role = "ctb"), person("Majid", "Djennad", role = "ctb"), person("Marco", "Enea", role = "ctb"), person("Alexios", "Ghalanos", role = "ctb"), person("Christos", "Argyropoulos", role = "ctb"), person("Almond", "Stocker", role = "ctb"), person("Jens", "Lichter", role = "ctb"), person("Stanislaus", "Stadlmann", role = "ctb") )

Description: 	 A set of distributions which can be used for modelling the response variables in Generalized Additive Models for Location Scale and Shape, Rigby and Stasinopoulos (2005), <doi:10.1111/j.1467-9876.2005.00510.x>. The distributions can be continuous, discrete or mixed distributions. Extra distributions can be created, by transforming, any continuous distribution defined on the real line, to a distribution defined on ranges 0 to infinity or 0 to 1, by using a ''log'' or a ''logit' transformation respectively.

License: 	 GPL-2 | GPL-3

URL: 	 http://www.gamlss.org/

Depends: 	 R (>= 2.15.0), MASS, graphics, stats, methods, grDevices

Repository: 	 CRAN

Author: 	 Mikis Stasinopoulos [aut, cre, cph], Robert Rigby [aut], Calliope Akantziliotou [ctb], Vlasios Voudouris [ctb], Gillian Heller [ctb], Fernanda De Bastiani [ctb], Raydonal Ospina [ctb], Nicoletta Motpan [ctb], Fiona McElduff [ctb], Majid Djennad [ctb], Marco Enea [ctb], Alexios Ghalanos [ctb], Christos Argyropoulos [ctb], Almond Stocker [ctb], Jens Lichter [ctb], Stanislaus Stadlmann [ctb]

Maintainer: 	 Mikis Stasinopoulos <d.stasinopoulos@londonmet.ac.uk>

Index of help topics:

BB                      Beta Binomial Distribution For Fitting a GAMLSS
                        Model
BCCG                    Box-Cox Cole and Green distribution (or Box-Cox
                        normal) for fitting a GAMLSS
BCPE                    Box-Cox Power Exponential distribution for
                        fitting a GAMLSS
BCT                     Box-Cox t distribution for fitting a GAMLSS
BE                      The beta distribution for fitting a GAMLSS
BEINF                   The beta inflated distribution for fitting a
                        GAMLSS
BEOI                    The one-inflated beta distribution for fitting
                        a GAMLSS
BEZI                    The zero-inflated beta distribution for fitting
                        a GAMLSS
BI                      Binomial distribution for fitting a GAMLSS
BNB                     Beta Negative Binomial distribution for fitting
                        a GAMLSS
DBI                     The Double binomial distribution
DBURR12                 The Discrete Burr type XII distribution for
                        fitting a GAMLSS model
DEL                     The Delaporte distribution for fitting a GAMLSS
                        model
DPO                     The Double Poisson distribution
EGB2                    The exponential generalized Beta type 2
                        distribution for fitting a GAMLSS
EXP                     Exponential distribution for fitting a GAMLSS
GA                      Gamma distribution for fitting a GAMLSS
GAF                     The Gamma distribution family
GB1                     The generalized Beta type 1 distribution for
                        fitting a GAMLSS
GB2                     The generalized Beta type 2 and generalized
                        Pareto distributions for fitting a GAMLSS
GEOM                    Geometric distribution for fitting a GAMLSS
                        model
GG                      Generalized Gamma distribution for fitting a
                        GAMLSS
GIG                     Generalized Inverse Gaussian distribution for
                        fitting a GAMLSS
GPO                     The generalised Poisson distribution
GT                      The generalized t distribution for fitting a
                        GAMLSS
GU                      The Gumbel distribution for fitting a GAMLSS
IG                      Inverse Gaussian distribution for fitting a
                        GAMLSS
IGAMMA                  Inverse Gamma distribution for fitting a GAMLSS
JSU                     The Johnson's Su distribution for fitting a
                        GAMLSS
JSUo                    The original Johnson's Su distribution for
                        fitting a GAMLSS
LG                      Logarithmic and zero adjusted logarithmic
                        distributions for fitting a GAMLSS model
LNO                     Log Normal distribution for fitting in GAMLSS
LO                      Logistic distribution for fitting a GAMLSS
LOGITNO                 Logit Normal distribution for fitting in GAMLSS
LQNO                    Normal distribution with a specific mean and
                        variance relationship for fitting a GAMLSS
                        model
MN3                     Multinomial distribution in GAMLSS
NBF                     Negative Binomial Family distribution for
                        fitting a GAMLSS
NBI                     Negative Binomial type I distribution for
                        fitting a GAMLSS
NBII                    Negative Binomial type II distribution for
                        fitting a GAMLSS
NET                     Normal Exponential t distribution (NET) for
                        fitting a GAMLSS
NO                      Normal distribution for fitting a GAMLSS
NO2                     Normal distribution (with variance as sigma
                        parameter) for fitting a GAMLSS
NOF                     Normal distribution family for fitting a GAMLSS
PARETO2                 Pareto distributions for fitting in GAMLSS
PE                      Power Exponential distribution for fitting a
                        GAMLSS
PIG                     The Poisson-inverse Gaussian distribution for
                        fitting a GAMLSS model
PO                      Poisson distribution for fitting a GAMLSS model
RG                      The Reverse Gumbel distribution for fitting a
                        GAMLSS
RGE                     Reverse generalized extreme family distribution
                        for fitting a GAMLSS
SEP                     The Skew Power exponential (SEP) distribution
                        for fitting a GAMLSS
SEP1                    The Skew Power exponential type 1-4
                        distribution for fitting a GAMLSS
SHASH                   The Sinh-Arcsinh (SHASH) distribution for
                        fitting a GAMLSS
SI                      The Sichel dustribution for fitting a GAMLSS
                        model
SICHEL                  The Sichel distribution for fitting a GAMLSS
                        model
SIMPLEX                 The simplex distribution for fitting a GAMLSS
SN1                     Skew Normal Type 1 distribution for fitting a
                        GAMLSS
SN2                     Skew Normal Type 2 distribution for fitting a
                        GAMLSS
ST1                     The skew t distributions, type 1 to 5
TF                      t family distribution for fitting a GAMLSS
WARING                  Waring distribution for fitting a GAMLSS model
WEI                     Weibull distribution for fitting a GAMLSS
WEI2                    A specific parameterization of the Weibull
                        distribution for fitting a GAMLSS
WEI3                    A specific parameterization of the Weibull
                        distribution for fitting a GAMLSS
YULE                    Yule distribution for fitting a GAMLSS model
ZABB                    Zero inflated and zero adjusted Binomial
                        distribution for fitting in GAMLSS
ZABI                    Zero inflated and zero adjusted Binomial
                        distribution for fitting in GAMLSS
ZAGA                    The zero adjusted Gamma distribution for
                        fitting a GAMLSS model
ZAIG                    The zero adjusted Inverse Gaussian distribution
                        for fitting a GAMLSS model
ZANBI                   Zero inflated and zero adjusted negative
                        binomial distributions for fitting a GAMLSS
                        model
ZAP                     Zero adjusted poisson distribution for fitting
                        a GAMLSS model
ZIP                     Zero inflated poisson distribution for fitting
                        a GAMLSS model
ZIP2                    Zero inflated poisson distribution for fitting
                        a GAMLSS model
ZIPF                    The zipf and zero adjusted zipf distributions
                        for fitting a GAMLSS model
checklink               Set the Right Link Function for Specified
                        Parameter and Distribution
count_1_31              A set of functions to plot gamlss.family
                        distributions
exGAUS                  The ex-Gaussian distribution
flexDist                Non-parametric pdf from limited information
                        data
gamlss.dist-package     Distributions for Generalized Additive Models
                        for Location Scale and Shape
gamlss.family           Family Objects for fitting a GAMLSS model
gen.Family              Functions to generate log and logit
                        distributions from existing continuous
                        gamlss.family distributions
hazardFun               Hazard functions for gamlss.family
                        distributions
make.link.gamlss        Create a Link for GAMLSS families
momentSK                Sample and theoretical Moment and Centile
                        Skewness and Kurtosis Functions

Author(s)

Mikis Stasinopoulos [aut, cre, cph], Robert Rigby [aut], Calliope Akantziliotou [ctb], Vlasios Voudouris [ctb], Gillian Heller [ctb], Fernanda De Bastiani [ctb], Raydonal Ospina [ctb], Nicoletta Motpan [ctb], Fiona McElduff [ctb], Majid Djennad [ctb], Marco Enea [ctb], Alexios Ghalanos [ctb], Christos Argyropoulos [ctb], Almond Stocker [ctb], Jens Lichter [ctb], Stanislaus Stadlmann [ctb]

Maintainer: Mikis Stasinopoulos <d.stasinopoulos@londonmet.ac.uk>

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family

Examples

# pdf plot
plot(function(y) dSICHEL(y, mu=10, sigma = 0.1 , nu=1 ), 
              from=0, to=30, n=30+1, type="h")
# cdf plot
PPP <- par(mfrow=c(2,1))
plot(function(y) pSICHEL(y, mu=10, sigma =0.1, nu=1 ), 
             from=0, to=30, n=30+1, type="h") # cdf
cdf<-pSICHEL(0:30, mu=10, sigma=0.1, nu=1) 
sfun1  <- stepfun(1:30, cdf, f = 0)
plot(sfun1, xlim=c(0,30), main="cdf(x)")
par(PPP)

[Package gamlss.dist version 5.1-4 Index]