gamlss.dist-package {gamlss.dist}	R Documentation

The GAMLSS distributions

Description

This package contains all distributions to be used for GAMLSS models. Each distributions has its probability function, d, its commutative probability function, p, the inverse of the commutative probability function, q, its random generation function, r, and also the gamlss.family generating function

Details

Package: 	 gamlss.dist

Type: 	 Package

Version: 	 1.5.0

Date: 	 2006-12-13

License: 	 GPL (version 2 or later)

This package is design to be used with the package gamlss but the d, p, q and r functions can be used separately.

Author(s)

Mikis Stasinopoulos <mikis.stasinopoulos@gamlss.org>, Bob Rigby <r.rigby@londonmet.ac.uk> with contributions from Calliope Akantziliotou and Raydonal Ospina <rospina@ime.usp.br>.

Maintainer: Mikis Stasinopoulos <mikis.stasinopoulos@gamlss.org>

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2003) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/)

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family

Examples

plot(function(y) dSICHEL(y, mu=10, sigma = 0.1 , nu=1 ), from=0, to=30, n=30+1, type="h") # pdf
# cdf plot
PPP <- par(mfrow=c(2,1))
plot(function(y) pSICHEL(y, mu=10, sigma =0.1, nu=1 ), from=0, to=30, n=30+1, type="h") # cdf
cdf<-pSICHEL(0:30, mu=10, sigma=0.1, nu=1) 
sfun1  <- stepfun(1:30, cdf, f = 0)
plot(sfun1, xlim=c(0,30), main="cdf(x)")
par(PPP)

[Package gamlss.dist version 5.0-5 Index]