gamlss.family {gamlss.dist}	R Documentation

Family Objects for fitting a GAMLSS model

Description

GAMLSS families are the current available distributions that can be fitted using the gamlss() function.

Usage

gamlss.family(object,...)
as.gamlss.family(object)
as.family(object)
## S3 method for class 'gamlss.family'
print(x,...)

Arguments

object
a gamlss family object e.g. BCT

x
a gamlss family object e.g. BCT

...
further arguments passed to or from other methods.

Details

There are several distributions available for the response variable in the gamlss function. The following table display their names and their abbreviations in R. Note that the different distributions can be fitted using their R abbreviations (and optionally excluding the brackets) i.e. family=BI(), family=BI are equivalent.

Distributions 	 R names 	 No of parameters

Beta 	 BE() 	 2

Beta Binomial 	 BB() 	 2

Beta negative binomial 	 BNB() 	 3

Beta one inflated 	 BEOI() 	 3

Beta zero inflated 	 BEZI() 	 3

Beta inflated 	 BEINF() 	 4

Binomial 	 BI() 	 1

Box-Cox Cole and Green 	 BCCG() 	 3

Box-Cox Power Exponential 	 BCPE() 	 4

Box-Cox-t 	 BCT() 	 4

Delaport 	 DEL() 	 3

Discrete Burr XII 	 DBURR12() 	 3

Double Poisson 	 DPO() 	 2

Double binomial 	 DBI() 	 2

Exponential 	 EXP() 	 1

Exponential Gaussian 	 exGAUS() 	 3

Exponential generalized Beta type 2 	 EGB2() 	 4

Gamma 	 GA() 	 2

Generalized Beta type 1 	 GB1() 	 4

Generalized Beta type 2 	 GB2() 	 4

Generalized Gamma 	 GG() 	 3

Generalized Inverse Gaussian 	 GIG() 	 3

Generalized t 	 GT() 	 4

Geometric 	 GEOM() 	 1

Geometric (original) 	 GEOMo() 	 1

Gumbel 	 GU() 	 2

Inverse Gamma 	 IGAMMA() 	 2

Inverse Gaussian 	 IG() 	 2

Johnson's SU 	 JSU() 	 4

Logarithmic 	 LG() 	 1

Logistic 	 LO() 	 2

Logit-Normal 	 LOGITNO() 	 2

log-Normal 	 LOGNO() 	 2

log-Normal (Box-Cox) 	 LNO() 	 3 (1 fixed)

Negative Binomial type I 	 NBI() 	 2

Negative Binomial type II 	 NBII() 	 2

Negative Binomial family 	 NBF() 	 3

Normal Exponential t 	 NET() 	 4 (2 fixed)

Normal 	 NO() 	 2

Normal Family 	 NOF() 	 3 (1 fixed)

Normal Linear Quadratic 	 LQNO() 	 2

Pareto type 2 	 PARETO2() 	 2

Pareto type 2 original 	 PARETO2o() 	 2

Power Exponential 	 PE() 	 3

Power Exponential type 2 	 PE2() 	 3

Poison 	 PO() 	 1

Poisson inverse Gaussian 	 PIG() 	 2

Reverse generalized extreme 	 RGE() 	 3

Reverse Gumbel 	 RG() 	 2

Skew Power Exponential type 1 	 SEP1() 	 4

Skew Power Exponential type 2 	 SEP2() 	 4

Skew Power Exponential type 3 	 SEP3() 	 4

Skew Power Exponential type 4 	 SEP4() 	 4

Shash 	 SHASH() 	 4

Shash original 	 SHASHo() 	 4

Shash original 2 	 SHASH() 	 4

Sichel (original) 	 SI() 	 3

Sichel (mu as the maen) 	 SICHEL() 	 3

Simplex 	 SIMPLEX() 	 2

Skew t type 1 	 ST1() 	 3

Skew t type 2 	 ST2() 	 3

Skew t type 3 	 ST3() 	 3

Skew t type 4 	 ST4() 	 3

Skew t type 5 	 ST5() 	 3

t-distribution 	 TF() 	 3

Waring 	 WARING() 	 1

Weibull 	 WEI() 	 2

Weibull(PH parameterization) 	 WEI2() 	 2

Weibull (mu as mean) 	 WEI3() 	 2

Yule 	 YULE() 	 1

Zero adjusted binomial 	 ZABI() 	 2

Zero adjusted beta neg. bin. 	 ZABNB() 	 4

Zero adjusted IG 	 ZAIG() 	 2

Zero adjusted logarithmic 	 ZALG() 	 2

Zero adjusted neg. bin. 	 ZANBI() 	 3

Zero adjusted poisson 	 ZAP() 	 2

Zero adjusted Sichel 	 ZASICHEL()	 4

Zero adjusted Zipf 	 ZAZIPF() 	 2

Zero inflated binomial 	 ZIBI() 	 2

Zero inflated beta neg. bin. 	 ZIBNB() 	 4

Zero inflated neg. bin. 	 ZINBI() 	 3

Zero inflated poisson 	 ZIP() 	 2

Zero inf. poiss.(mu as mean) 	 ZIP2() 	 2

Zero inflated PIG 	 ZIPIG() 	 3

Zero inflated Sichel 	 ZISICHEL()	 4

Zipf 	 ZIPF() 	 1

Note that some of the distributions are in the package gamlss.dist. The parameters of the distributions are in order, mu for location, sigma for scale (or dispersion), and nu and tau for shape. More specifically for the BCCG family mu is the median, sigma approximately the coefficient of variation, and nu the skewness parameter. The parameters for BCPE distribution have the same interpretation with the extra fourth parameter tau modelling the kurtosis of the distribution. The parameters for BCT have the same interpretation except that sigma*((tau/(tau-2))^0.5) is approximately the coefficient of variation.

All of the distribution in the above list are also provided with the corresponding d, p, q and r functions for density (pdf), distribution function (cdf), quantile function and random generation function respectively, (see individual distribution for details).

Value

The above GAMLSS families return an object which is of type gamlss.family. This object is used to define the family in the gamlss() fit.

Note

More distributions will be documented in later GAMLSS releases. Further user defined distributions can be incorporate relatively easy, see, for example, the help documentation accompanying the gamlss library.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Rigby, R. A., Stasinopoulos, D. M., Heller, G. Z., and De Bastiani, F. (2019) Distributions for modeling location, scale, and shape: Using GAMLSS in R, Chapman and Hall/CRC. An older version can be found in http://www.gamlss.com/.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

BE,BB,BEINF,BI,LNO,BCT, BCPE,BCCG, GA,GU,JSU,IG,LO, NBI,NBII,NO,PE,PO, RG,PIG,TF,WEI,WEI2, ZIP

Examples

 normal<-NO(mu.link="log", sigma.link="log")
 normal

[Package gamlss.dist version 5.1-5 Index]