BI(gamlss.dist)	R Documentation

Binomial distribution for fitting a GAMLSS

Description

The BI() function defines the binomial distribution, a one parameter family distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dBI, pBI, qBI and rBI define the density, distribution function, quantile function and random generation for the binomial, BI(), distribution.

Usage


BI(mu.link = "logit")
dBI(x, bd = 1, mu = 0.5, log = FALSE)
pBI(q, bd = 1, mu = 0.5, lower.tail = TRUE, log.p = FALSE)
qBI(p, bd = 1, mu = 0.5, lower.tail = TRUE, log.p = FALSE)
rBI(n, bd = 1, mu = 0.5)

Arguments

mu.link 	 Defines the mu.link, with "logit" link as the default for the mu parameter. Other links are "probit" and "cloglog"'(complementary log-log)

x 	 vector of (non-negative integer) quantiles

mu 	 vector of positive probabilities

bd 	 vector of binomial denominators

p 	 vector of probabilities

q 	 vector of quantiles

n 	 number of random values to return

log, log.p 	 logical; if TRUE, probabilities p are given as log(p)

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

Details

Definition file for binomial distribution.

f(y|mu)=(Gamma(n+1)*Gamma(y+1)/Gamma(n-y+1))* mu^y *(1-mu)^(n-y)

for y=0,1,2,...,n and 0<μ< 1.

Value

returns a gamlss.family object which can be used to fit a binomial distribution in the gamlss() function.

Note

The response variable should be a matrix containing two columns, the first with the count of successes and the second with the count of failures. The parameter mu represents a probability parameter with limits 0 < mu <1. n*mu is the mean of the distribution where n is the binomial denominator.

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

Examples


 BI()# gives information about the default links for the Binomial distribution 
# data(aep)   
# library(gamlss)
# h<-gamlss(y~ward+loglos+year, family=BI, data=aep)  
# plot of the binomial distribution
curve(dBI(x, mu = .5, bd=10), from=0, to=10, n=10+1, type="h")
tN <- table(Ni <- rBI(1000, mu=.2, bd=10))
r <- barplot(tN, col='lightblue')

[Package gamlss.dist version 3.0-0 Index]