EXP(gamlss.dist)	R Documentation

Exponential distribution for fitting a GAMLSS

Description

The function EXP defines the exponential distribution, a one parameter distribution for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The mu parameter represents the mean of the distribution. The functions dEXP, pEXP, qEXP and rEXP define the density, distribution function, quantile function and random generation for the specific parameterization of the exponential distribution defined by function EXP.

Usage


EXP(mu.link ="log")
dEXP(x, mu = 1, log = FALSE)
pEXP(q, mu = 1, lower.tail = TRUE, log.p = FALSE)
qEXP(p, mu = 1, lower.tail = TRUE, log.p = FALSE)
rEXP(n, mu = 1)

Arguments

mu.link 	 Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse" and "identity"

x,q 	 vector of quantiles

mu 	 vector of location parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

The specific parameterization of the exponential distribution used in EXP is

f(y|mu)=1/mu*exp(-y/mu)

, for y>0, mu>0.

Value

EXP() returns a gamlss.family object which can be used to fit an exponential distribution in the gamlss() function. dEXP() gives the density, pEXP() gives the distribution function, qEXP() gives the quantile function, and rEXP() generates random deviates.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Nicoleta Motpan

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

Examples


y<-rEXP(1000,mu=1) # generates 1000 random observations 
hist(y)
# library(gamlss)
# histDist(y, family=EXP) 

[Package gamlss.dist version 3.0-0 Index]