GEOM {gamlss.dist}	R Documentation

Geometric distribution for fitting a GAMLSS model

Description

The functions GEOMo() and GEOM() define two parametrizations of the geometric distribution. The geometric distribution is a one parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The mean of GEOM() is equal to the parameter mu. The functions dGEOM, pGEOM, qGEOM and rGEOM define the density, distribution function, quantile function and random generation for the GEOM parameterization of the Geometric distribution.

Usage

GEOM(mu.link = "log")
dGEOM(x, mu = 2, log = FALSE)
pGEOM(q, mu = 2, lower.tail = TRUE, log.p = FALSE)
qGEOM(p, mu = 2, lower.tail = TRUE, log.p = FALSE)
rGEOM(n, mu = 2)
GEOMo(mu.link = "logit")
dGEOMo(x, mu = 0.5, log = FALSE)
pGEOMo(q, mu = 0.5, lower.tail = TRUE, log.p = FALSE)
qGEOMo(p, mu = 0.5, lower.tail = TRUE, log.p = FALSE)
rGEOMo(n, mu = 0.5)

Arguments

mu.link
Defines the mu.link, with log link as the default for the mu parameter

x, q
vector of quantiles

mu
vector of location parameter values

log, log.p
logical; if TRUE, probabilities p are given as log(p)

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise P[X > x]

p
vector of probabilities

n
number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the original geometric distribution in the function GE is

f(y|μ) = (1-μ)^y μ

for y>=0 and μ>0.

The parameterization of the geometric distribution in the function GEOM is

f(y|μ) = μ^y/(μ+1)^{y+1}

where for y>=0 and μ>0.

Value

returns a gamlss.family object which can be used to fit a Geometric distribution in the gamlss() function.

Author(s)

Fiona McElduff, Bob Rigby and Mikis Stasinopoulos.

References

Johnson, N. L., Kemp, A. W., and Kotz, S. (2005). Univariate discrete distributions. Wiley.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.org/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Stasinopoulos D. M., Rigby R.A., Heller G., Voudouris V., and De Bastiani F., (2017) Flexible Regression and Smoothing: Using GAMLSS in R, Chapman and Hall/CRC.

See Also

gamlss.family

Examples

par(mfrow=c(2,2))
y<-seq(0,20,1)
plot(y, dGEOM(y), type="h")
q <- seq(0, 20, 1)
plot(q, pGEOM(q), type="h")
p<-seq(0.0001,0.999,0.05)
plot(p , qGEOM(p), type="s")
dat <- rGEOM(100)
hist(dat)
#summary(gamlss(dat~1, family=GEOM))
par(mfrow=c(2,2))
y<-seq(0,20,1)
plot(y, dGEOMo(y), type="h")
q <- seq(0, 20, 1)
plot(q, pGEOMo(q), type="h")
p<-seq(0.0001,0.999,0.05)
plot(p , qGEOMo(p), type="s")
dat <- rGEOMo(100)
hist(dat)
#summary(gamlss(dat~1, family="GE"))

[Package gamlss.dist version 5.0-5 Index]