GIG(gamlss.dist)	R Documentation

Generalized Inverse Gaussian distribution for fitting a GAMLSS

Description

The function GIG defines the generalized inverse gaussian distribution, a three parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions DIG, pGIG, GIG and rGIG define the density, distribution function, quantile function and random generation for the specific parameterization of the generalized inverse gaussian distribution defined by function GIG.

Usage


GIG(mu.link = "log", sigma.link = "log", 
                       nu.link = "identity")
dGIG(x, mu=1, sigma=1, nu=1,  
                      log = FALSE)
pGIG(q, mu=1, sigma=1, nu=1,  lower.tail = TRUE, 
                     log.p = FALSE)
qGIG(p, mu=1, sigma=1, nu=1,  lower.tail = TRUE, 
                     log.p = FALSE)
rGIG(n, mu=1, sigma=1, nu=1, ...)

Arguments

mu.link 	 Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse" and "identity"

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter, other links are "inverse" and "identity"

nu.link 	 Defines the nu.link, with "identity" link as the default for the nu parameter, other links are "inverse" and "log"

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of shape parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

... 	 for extra arguments

Details

The specific parameterization of the generalized inverse gaussian distribution used in GIG is f(y|mu,sigma,nu)=((c/mu)^nu)*(y^(nu-1))/(2*besselK(1/sigma,nu))*exp(-1/(2*sigma)*(c*y/mu+mu/(c*y))) where c = besselK(1/sigma,nu+1)/besselK(1/sigma,nu), for y>0, mu>0, sigma>0 and -Inf>nu>Inf.

Value

GIG() returns a gamlss.family object which can be used to fit a generalized inverse gaussian distribution in the gamlss() function. DIG() gives the density, pGIG() gives the distribution function, GIG() gives the quantile function, and rGIG() generates random deviates.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Nicoleta Motpan

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

Jorgensen B. (1982) Statistical properties of the generalized inverse Gaussian distribution, Series: Lecture notes in statistics; 9, New York : Springer-Verlag.

See Also

gamlss, gamlss.family, GI

Examples


y<-rGIG(100,mu=1,sigma=1, nu=-0.5) # generates 1000 random observations 
hist(y)
# library(gamlss)
# histDist(y, family=GIG) 

[Package gamlss.dist version 3.0-0 Index]