GT(gamlss.dist)	R Documentation

The generalized t distribution for fitting a GAMLSS

Description

This function defines the generalized t distribution, a four parameter distribution, for a gamlss.family object to be used for a GAMLSS fitting using the function gamlss(). The functions dGT, pGT, qGT and rGT define the density, distribution function, quantile function and random generation for the generalized t distribution.

Usage


GT(mu.link = "identity", sigma.link = "log", nu.link = "log", 
   tau.link = "log")
dGT(x, mu = 0, sigma = 1, nu = 3, tau = 1.5, log = FALSE)
pGT(q, mu = 0, sigma = 1, nu = 3, tau = 1.5, lower.tail = TRUE, 
   log.p = FALSE)
qGT(p, mu = 0, sigma = 1, nu = 3, tau = 1.5, lower.tail = TRUE, 
   log.p = FALSE)
rGT(n, mu = 0, sigma = 1, nu = 3, tau = 1.5)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter.

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter.

nu.link 	 Defines the nu.link, with "log" link as the default for the nu parameter.

tau.link 	 Defines the tau.link, with "log" link as the default for the tau parameter.

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of skewness nu parameter values

tau 	 vector of kurtosis tau parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

The probability density function of the generalized t distribution, (GT), , is defined as

f(y|mu,sigma,nu,tau)=

where 0<y<0, z=(y-mu)/sigma mu=(-Inf,+Inf), sigma>0, nu>0 and tau>0.

Value

GT() returns a gamlss.family object which can be used to fit the GT distribution in the gamlss() function. dGT() gives the density, pGT() gives the distribution function, qGT() gives the quantile function, and rGT() generates random deviates.

Warning

The qSHASH and rSHASH are slow since they are relying on golden section for finding the quantiles

Author(s)

Bob Rigby r.rigby@londonmet.ac.uk and Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R. A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, JSU, BCT

Examples


GT()   # 
y<- rGT(200, mu=5, sigma=1, nu=1, tau=4)
hist(y)
curve(dGT(x, mu=5 ,sigma=2,nu=1, tau=4), -2, 11, main = "The GT  density mu=5 ,sigma=1, nu=1, tau=4")
# library(gamlss)
# m1<-gamlss(y~1, family=GT) 

[Package gamlss.dist version 3.0-0 Index]