NBII(gamlss.dist)	R Documentation

Negative Binomial type II distribution for fitting a GAMLSS

Description

The NBII() function defines the Negative Binomial type II distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dNBII, pNBII, qNBII and rNBII define the density, distribution function, quantile function and random generation for the Negative Binomial type II, NBII(), distribution.

Usage


NBII(mu.link = "log", sigma.link = "log")
dNBII(x, mu = 1, sigma = 1, log = FALSE)
pNBII(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qNBII(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rNBII(n, mu = 1, sigma = 1)

Arguments

mu.link 	 Defines the mu.link, with "log" link as the default for the mu parameter

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter

x 	 vector of (non-negative integer) quantiles

mu 	 vector of positive means

sigma 	 vector of positive despersion parameter

p 	 vector of probabilities

q 	 vector of quantiles

n 	 number of random values to return

log, log.p 	 logical; if TRUE, probabilities p are given as log(p)

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

Details

Definition file for Negative Binomial type II distribution.

f(y|mu,sigma)=Gamma(y+(mu/sigma))*sigma^y/(Gamma(mu/sigma)*Gamma(y+1)*(1+sigma)^(y+(mu/sigma)))

for y=0,1,2, ...,Inf, μ>0 and σ>0. This parameterization was used by Evans (1953) and also by Johnson et al. (1993) p 200.

Value

returns a gamlss.family object which can be used to fit a Negative Binomial type II distribution in the gamlss() function.

Note

mu is the mean and ((1+sigma)*mu)**0.5 is the standard deviation of the Negative Binomial type II distribution, so sigma is a dispersion parameter

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Calliope Akantziliotou

References

Evans, D. A. (1953). Experimental evidence concerning contagious distributions in ecology. Biometrika, 40: 186-211.

Johnson, N. L., Kotz, S. and Kemp, A. W. (1993). Univariate Discrete Distributions, 2nd edn. Wiley, New York.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, NBI, PIG, SI

Examples


NBII()  # gives information about the default links for the Negative Binomial type II distribution  
# plotting the distribution
plot(function(y) dNBII(y, mu = 10, sigma = 0.5 ), from=0, to=40, n=40+1, type="h")
# creating random variables and plot them 
tN <- table(Ni <- rNBII(1000, mu=5, sigma=0.5))
r <- barplot(tN, col='lightblue')
# library(gamlss)
# data(aids)
# h<-gamlss(y~cs(x,df=7)+qrt, family=NBII, data=aids) # fits a model 
# plot(h)
# pdf.plot(family=NBII, mu=10, sigma=0.5, min=0, max=40, step=1)

[Package gamlss.dist version 3.0-0 Index]