NET(gamlss.dist)	R Documentation

Normal Exponential t distribution (NET) for fitting a GAMLSS

Description

This function defines the Power Exponential t distribution (NET), a four parameter distribution, for a gamlss.family object to be used for a GAMLSS fitting using the function gamlss(). The functions dNET, pNET define the density and distribution function the NET distribution.

Usage


NET(mu.link = "identity", sigma.link = "log")
pNET(q, mu = 5, sigma = 0.1, nu = 1, tau = 2)
dNET(x, mu = 0, sigma = 1, nu = 1.5, tau = 2, log = FALSE)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter. Other links are "inverse", "log" and "own"

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter. Other links are "inverse", "identity" and "own"

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of nu parameter values

tau 	 vector of tau parameter values

log 	 logical; if TRUE, probabilities p are given as log(p).

Details

The NET distribution was introduced by Rigby and Stasinopoulos (1994) as a robust distribution for a response variable with heavier tails than the normal. The NET distribution is the abbreviation of the Normal Exponential Student t distribution. The NET distribution is a four parameter continuous distribution, although in the GAMLSS implementation only the two parameters, mu and sigma, of the distribution are modelled with nu and tau fixed. The distribution takes its names because it is normal up to nu, Exponential from nu to tau (hence abs(nu)<=abs(tau)) and Student-t with nu*tau-1 degrees of freedom after tau. Maximum likelihood estimator of the third and forth parameter can be obtained, using the GAMLSS functions, find.hyper or prof.dev.

Value

NET() returns a gamlss.family object which can be used to fit a Box Cox Power Exponential distribution in the gamlss() function. dNET() gives the density, pNET() gives the distribution function.

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos, D. M. (1994), Robust fitting of an additive model for variance heterogeneity, COMPSTAT : Proceedings in Computational Statistics, editors:R. Dutter and W. Grossmann, pp 263-268, Physica, Heidelberg.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R. A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, BCPE

Examples


NET()   # 
data(abdom)
plot(function(x)dNET(x, mu=0,sigma=1,nu=2, tau=3), -5, 5)
plot(function(x)pNET(x, mu=0,sigma=1,nu=2, tau=3), -5, 5) 
# fit NET with nu=1 and tau=3
# library(gamlss)
#h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=NET, 
#        data=abdom, nu.start=2, tau.start=3) 
#plot(h)

[Package gamlss.dist version 3.0-0 Index]