NOF(gamlss.dist)	R Documentation

Normal distribution family for fitting a GAMLSS

Description

The function NOF() defines a normal distribution family, which has three parameters. The distribution can be used in a GAMLSS fitting using the function gamlss(). The mean of NOF is equal to mu. The variance is equal to sigma^2*mu^nu so the standard deviation is sigma*mu^(nu/2). The function is design for cases where the variance is proportional to a power of the mean. The functions dNOF, pNOF, qNOF and rNOF define the density, distribution function, quantile function and random generation for the NOF parametrization of the normal distribution family.

Usage


NOF(mu.link = "identity", sigma.link = "log", nu.link = "identity")
dNOF(x, mu = 0, sigma = 1, nu = 0, log = FALSE)
pNOF(q, mu = 0, sigma = 1, nu = 0, lower.tail = TRUE, log.p = FALSE)
qNOF(p, mu = 0, sigma = 1, nu = 0, lower.tail = TRUE, log.p = FALSE)
rNOF(n, mu = 0, sigma = 1, nu = 0)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter

nu.link 	 Defines the nu.link with "identity" link as the default for the nu parameter

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of power parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parametrization of the normal distribution given in the function NOF() is

f(y|mu,sigma,nu)=(1/(sqrt(2*pi)*sigma*mu^(nu/2)))* exp(-0.5*((y-mu)^2/sigma^2*mu^nu))

for y=(-Inf,Inf), μ=(-Inf,Inf), σ>0 and nu=(-Inf,+Inf).

Value

returns a gamlss.family object which can be used to fit a normal distribution family in the gamlss() function.

Note

For the function NOF(), mu is the mean and sigma*mu^(nu/2) is the standard deviation of the normal distribution family. The NOF is design for fitting regression type models where the variance is proportional to a power ofthe mean. Models of this type are related to the "pseudo likelihood" models of Carroll and Rubert (1987) but here a proper likelihood is miximised.

Note that because the high correlation between the sigma and the nu parameter the mixed() method should be used in the fitting.

Author(s)

Mikis Stasinopoulos, Bob Rigby and Calliope Akantziliotou

References

Davidian, M. and Carroll, R. J. (1987), Variance Function Estimation, Journal of the American Statistical Association, Vol. 82, pp. 1079-1091

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, NO, NO2

Examples


NOF()# gives information about the default links for the normal distribution family
# library(gamlss)
#data(abdom)        
## the normal distribution fit with constant sigma
#m1<-gamlss(y~poly(x,2), sigma.fo=~1, family=NO, data=abdom)
## the normal family fit with variance proportional to mu
#m2<-gamlss(y~poly(x,2), sigma.fo=~1, family=NOF, data=abdom, method=mixed(1,20))
## a nornal distribution fit with variance as a function of x
#m3 <-gamlss(y~poly(x,2), sigma.fo=~x,   family=NO, data=abdom, method=mixed(1,20)) 
#GAIC(m1,m2,m3)

[Package gamlss.dist version 3.0-0 Index]