RGE(gamlss.dist)	R Documentation

Reverse generalized extreme family distribution for fitting a GAMLSS

Description

The function RGE defines the reverse generalized extreme family distribution, a three parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dRGE, pRGE, qRGE and rRGE define the density, distribution function, quantile function and random generation for the specific parameterization of the reverse generalized extreme distribution given in details below.

Usage


RGE(mu.link = "identity", sigma.link = "log", nu.link = "log")
dRGE(x, mu = 1, sigma = 0.1, nu = 1, log = FALSE)
pRGE(q, mu = 1, sigma = 0.1, nu = 1, lower.tail = TRUE, log.p = FALSE)
qRGE(p, mu = 1, sigma = 0.1, nu = 1, lower.tail = TRUE, log.p = FALSE) 
rRGE(n, mu = 1, sigma = 0.1, nu = 1)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter

nu.link 	 Defines the nu.link, with "log" link as the default for the nu parameter

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of the shape parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

Definition file for reverse generalized extreme family distribution.

The probability density function of the generalized extreme value distribution is obtained from Johnson et al. (1995), Volume 2, p76, equation (22.184) [where (xi,theta,gamma)->(mu,sigma, nu)].

The probability density function of the reverse generalized extreme value distribution is then obtained by replacing y by -y and μ by -μ.

Hence the probability density function of the reverse generalized extreme value distribution with nu>0 is given by

f(y|mu,sigma,nu)=(1/sigma)(1+(nu*(y-mu))/(sigma))^(1/(nu-1))*S1(y|mu,sigma,nu)

for

μ-frac{σ}{nu}<y<infty

where

S1(y|mu,sigma,nu)=exp(-[1+(nu*(y-mu))/(sigms)]^(1/nu))

and where -infty<μ<y+frac{σ}{nu}, σ>0 and nu>0. Note that only the case nu>0 is allowed here. The reverse generalized extreme value distribution is denoted as RGE(μ,σ,nu) or as Reverse Generalized.Extreme.Family(μ,σ,nu).

Note the the above distribution is a reparameterization of the three parameter Weibull distribution given by

f(y|mu,sigma,nu)=(a3/a2)*((y-a1)/a2)^(a3-1)exp(-((y-a1)/a2)^a3)

given by setting a1=mu-(sigma/nu), a2=sigma/nu, 1/nu.

Value

RGE() returns a gamlss.family object which can be used to fit a reverse generalized extreme distribution in the gamlss() function. dRGE() gives the density, pRGE() gives the distribution function, qRGE() gives the quantile function, and rRGE() generates random deviates.

Note

This distribution is very difficult to fit because the y values depends on the parameter values. The RS() and CG() algorithms are not appropriate for this type of problem.

Author(s)

Bob Rigby r.rigby@londonmet.ac.uk, Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk and Kalliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

Examples


RGE()# gives information about the default links for the reverse generalized extreme family distribution 
newdata<-rRGE(100,mu=0,sigma=1,nu=5) # generates 100 random observations
# library(gamlss)
# gamlss(newdata~1, family=RGE, method=mixed(5,50)) # difficult to converse 

[Package gamlss.dist version 3.0-0 Index]