SEP(gamlss.dist)	R Documentation

The Skew Power exponential (SEP) distribution for fitting a GAMLSS

Description

This function defines the Skew Power exponential (SEP) distribution, a four parameter distribution, for a gamlss.family object to be used for a GAMLSS fitting using the function gamlss(). The functions dSEP, pSEP, qSEP and rSEP define the density, distribution function, quantile function and random generation for the Skew Power exponential (SEP) distribution.

Usage


SEP(mu.link = "identity", sigma.link = "log", nu.link = "identity", 
    tau.link = "log")
dSEP(x, mu = 0, sigma = 1, nu = 0, tau = 2, log = FALSE)
pSEP(q, mu = 0, sigma = 1, nu = 0, tau = 2, lower.tail = TRUE, 
     log.p = FALSE)
qSEP(p, mu = 0, sigma = 1, nu = 0, tau = 2, lower.tail = TRUE, 
     log.p = FALSE, lower.limit = mu - 5 * sigma, 
     upper.limit = mu + 5 * sigma)
rSEP(n, mu = 0, sigma = 1, nu = 0, tau = 2)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter. Other links are "1/mu^2" and "log"

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter. Other links are "inverse" and "identity"

nu.link 	 Defines the nu.link, with "identity" link as the default for the nu parameter. Other links are "1/nu^2" and "log"

tau.link 	 Defines the tau.link, with "log" link as the default for the tau parameter. Other links are "1/tau^2", and "identity

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of skewness nu parameter values

tau 	 vector of kurtosis tau parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

lower.limit 	 lower limit for the golden search to find quantiles from probabilities

upper.limit 	 upper limit for the golden search to find quantiles from probabilities

Details

The probability density function of the Skew Power exponential distribution, (SEP), is defined as

f(y|mu,sigma,nu,tau)=(z/sigma)*pnorm(w)*dPE(z,0,1,tau)

for 0<y<0, mu=(-Inf,+Inf), sigma>0, nu=(-Inf,+Inf) and tau>0. where z=(y-mu)/(sigma), w=sign(z)|z|^(t/2) *nu*sqrt(2/tau) and dPE(z,0,1,tau) is the pdf of an Exponential Power distribution.

Value

SEP() returns a gamlss.family object which can be used to fit the SEP distribution in the gamlss() function. dSEP() gives the density, pSEP() gives the distribution function, qSEP() gives the quantile function, and rSEP() generates random deviates.

Warning

The qSEP and rSEP are slow since they are relying on golden section for finding the quantiles

Author(s)

Bob Rigby r.rigby@londonmet.ac.uk and Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk

References

Diciccio, T. J. and Mondi A. C. (2004). Inferential Aspects of the Skew Exponential Power distribution., JASA, 99, 439-450.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M. Rigby R. A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, JSU, BCT

Examples


SEP()   # 
plot(function(x)dSEP(x, mu=0,sigma=1, nu=1, tau=2), -5, 5, 
 main = "The SEP  density mu=0,sigma=1,nu=1, tau=2")
plot(function(x) pSEP(x, mu=0,sigma=1,nu=1, tau=2), -5, 5, 
 main = "The BCPE  cdf mu=0, sigma=1, nu=1, tau=2")
dat <- rSEP(100,mu=10,sigma=1,nu=-1,tau=1.5)
# library(gamlss)
# gamlss(dat~1,family=SEP, control=gamlss.control(n.cyc=30))

[Package gamlss.dist version 3.0-0 Index]