TF(gamlss.dist)	R Documentation

t family distribution for fitting a GAMLSS

Description

The function TF defines the t-family distribution, a three parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dTF, pTF, qTF and rTF define the density, distribution function, quantile function and random generation for the specific parameterization of the t distribution given in details below, with mean equal to mu and standard deviation equal to sigma*(nu/(nu-2))^0.5 with the degrees of freedom nu

Usage


TF(mu.link = "identity", sigma.link = "log", nu.link = "log")
dTF(x, mu = 0, sigma = 1, nu = 10, log = FALSE)
pTF(q, mu = 0, sigma = 1, nu = 10, lower.tail = TRUE, log.p = FALSE)
qTF(p, mu = 0, sigma = 1, nu = 10, lower.tail = TRUE, log.p = FALSE) 
rTF(n, mu = 0, sigma = 1, nu = 10)

Arguments

mu.link 	 Defines the mu.link, with "identity" link as the default for the mu parameter

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter

nu.link 	 Defines the nu.link, with "log" link as the default for the nu parameter

x,q 	 vector of quantiles

mu 	 vector of location parameter values

sigma 	 vector of scale parameter values

nu 	 vector of the degrees of freedom parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

Definition file for t family distribution.

f(y|mu,sigma,nu)=((Gamma((nu+1)/2)/(sigma*Gamma(1/2)*Gamma(nu/2))*nu^0.5 )*(1+(y-mu)^2/(nu*sigma^2))^(-(nu+1)/2)

y=(-Inf,+Inf), μ=(-Inf,+Inf), σ>0 and nu>0. Note that z=(y-mu)/sigma has a standard t distribution with degrees of freedom nu.

Value

TF() returns a gamlss.family object which can be used to fit a t distribution in the gamlss() function. dTF() gives the density, pTF() gives the distribution function, qTF() gives the quantile function, and rTF() generates random deviates. The latest functions are based on the equivalent R functions for gamma distribution.

Note

mu is the mean and sigma*(nu/(nu-2))^0.5 is the standard deviation of the t family distribution. nu>0 is a positive real valued parameter.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Kalliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

Examples


TF()# gives information about the default links for the t-family distribution 
# library(gamlss)
#data(abdom)
#h<-gamlss(y~cs(x,df=3), sigma.formula=~cs(x,1), family=TF, data=abdom) # fits 
#plot(h)
newdata<-rTF(1000,mu=0,sigma=1,nu=5) # generates 1000 random observations
hist(newdata) 

[Package gamlss.dist version 3.0-0 Index]