WEI2(gamlss.dist)	R Documentation

A specific parameterization of the Weibull distribution for fitting a GAMLSS

Description

The function WEI2 can be used to define the Weibull distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). This is the parameterization of the Weibull distribution usually used in proportional hazard models and is defined in details below. [Note that the GAMLSS function WEI uses a different parameterization for fitting the Weibull distribution.] The functions dWEI2, pWEI2, qWEI2 and rWEI2 define the density, distribution function, quantile function and random generation for the specific parameterization of the Weibull distribution.

Usage


WEI2(mu.link = "log", sigma.link = "log")
dWEI2(x, mu = 1, sigma = 1, log = FALSE)
pWEI2(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qWEI2(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rWEI2(n, mu = 1, sigma = 1)

Arguments

mu.link 	 Defines the mu.link, with "log" link as the default for the mu parameter, other links are "inverse" and "identity"

sigma.link 	 Defines the sigma.link, with "log" link as the default for the sigma parameter, other link is the "inverse" and "identity"

x,q 	 vector of quantiles

mu 	 vector of the mu parameter values

sigma 	 vector of sigma parameter values

log, log.p 	 logical; if TRUE, probabilities p are given as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required

Details

The parameterization of the function WEI2 is given by

f(y|mu,sigma)= sigma*mu*y^(sigma-1)exp(-mu*y^sigma)

for y>0, μ>0 and σ>0. The GAMLSS functions dWEI2, pWEI2, qWEI2, and rWEI2 can be used to provide the pdf, the cdf, the quantiles and random generated numbers for the Weibull distribution with argument mu, and sigma. [See the GAMLSS function WEI for a different parameterization of the Weibull.]

Value

WEI2() returns a gamlss.family object which can be used to fit a Weibull distribution in the gamlss() function. dWEI2() gives the density, pWEI2() gives the distribution function, qWEI2() gives the quantile function, and rWEI2() generates random deviates. The latest functions are based on the equivalent R functions for Weibull distribution.

Warning

In WEI2 the estimated parameters mu and sigma can be highly correlated so it is advisable to use the CG() method for fitting [as the RS() method can be veru slow in this situation.]

Note

The mean in WEI2 is given by mu^(-1/sigma)*gamma((1/sigma)+1) and the variance mu^{-2/sigma}*(gamma((2/sigma)+1)-gamma((1/sigma)+1)^2)

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, WEI,WEI3,

Examples


WEI2()
dat<-rWEI(100, mu=.1, sigma=2)
hist(dat)
# library(gamlss)
# gamlss(dat~1, family=WEI2, method=CG())


[Package gamlss.dist version 3.0-0 Index]