ZIP2(gamlss.dist)	R Documentation

Zero inflated poisson distribution for fitting a GAMLSS model

Description

The function ZIP2 defines the zero inflated Poisson type 2 distribution, a two parameter distribution, for a gamlss.family object to be used in GAMLSS fitting using the function gamlss(). The functions dZIP2, pZIP2, qZIP2 and rZIP2 define the density, distribution function, quantile function and random generation for the inflated poisson, ZIP2(), distribution. The ZIP2 is a different parameterization of the ZIP distribution. In the ZIP2 the mu is the mean of the distribution.

Usage


ZIP2(mu.link = "log", sigma.link = "logit")
dZIP2(x, mu = 5, sigma = 0.1, log = FALSE)
pZIP2(q, mu = 5, sigma = 0.1, lower.tail = TRUE, log.p = FALSE)
qZIP2(p, mu = 5, sigma = 0.1, lower.tail = TRUE, log.p = FALSE)
rZIP2(n, mu = 5, sigma = 0.1)

Arguments

mu.link 	 defines the mu.link, with "log" link as the default for the mu parameter

sigma.link 	 defines the sigma.link, with "logit" link as the default for the sigma parameter which in this case is the probability at zero. Other links are "probit" and "cloglog"'(complementary log-log)

x 	 vector of (non-negative integer) quantiles

mu 	 vector of positive means

sigma 	 vector of probabilities at zero

p 	 vector of probabilities

q 	 vector of quantiles

n 	 number of random values to return

log, log.p 	 logical; if TRUE, probabilities p are given as log(p)

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

Details

Let Y=0 with probability σ and Po(mu/(1-sigma)) with probability (1-σ) then Y has a Zero inflated Poisson type 2 distribution given by

sigma+(1-sigma)e^(-(mu/(1-sigma))) if y=0

f(y)=(1-sigma)exp(-(mu/(1-sigma)))* (mu/(1-sigma))^y/y! if y=0,1,2,...

The mean of the distribution in this parameterization is mu.

Value

returns a gamlss.family object which can be used to fit a zero inflated poisson distribution in the gamlss() function.

Author(s)

Bob Rigby, Gillian Heller and Mikis Stasinopoulos

References

Lambert, D. (1992), Zero-inflated Poisson Regression with an application to defects in Manufacturing, Technometrics, 34, pp 1-14.

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family, ZIP

Examples


ZIP2()# gives information about the default links for the normal distribution
# creating data and plotting them 
dat<-rZIP2(1000, mu=5, sigma=.1)
r <- barplot(table(dat), col='lightblue')
# fit the disteibution
# library(gamlss) 
# mod1<-gamlss(dat~1, family=ZIP2)# fits a constant for mu and sigma 
# fitted(mod1)[1]
# fitted(mod1,"sigma")[1]

[Package gamlss.dist version 3.0-0 Index]