checklink(gamlss.dist)	R Documentation

Set the Right Link Function for Specified Parameter and Distribution

Description

This function is used within the distribution family specification of a GAMLSS model to define the right link for each of the parameters of the distribution. This function should not be called by the user unless he/she specify a new distribution family or wishes to change existing link functions in the parameters.

Usage


checklink(which.link = NULL, which.dist = NULL, link = NULL, link.List = NULL)

Arguments

which.link 	 which parameter link e.g. which.link="mu.link"

which.dist 	 which distribution family e.g. which.dist="Cole.Green"

link 	 a repetition of which.link e.g. link=substitute(mu.link)

link.List 	 what link function are required e.g. link.List=c("inverse", "log", "identity")

Value

Defines the right link for each parameter

Author(s)

Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss, gamlss.family

[Package gamlss.dist version 3.0-0 Index]