gamlss.family(gamlss.dist)	R Documentation

Family Objects for fitting a GAMLSS model

Description

GAMLSS families are the current available distributions that can be fitted using the gamlss() function.

Usage


gamlss.family(object,...)
as.gamlss.family(object)
as.family(object)
## S3 method for class 'gamlss.family':
print(x,...)
gamlss.family.default(object,...)

Arguments

object 	 a gamlss family object e.g. BCT

x 	 a gamlss family object e.g. BCT

... 	 further arguments passed to or from other methods.

Details

There are several distributions available for the response variable in the gamlss function. The following table display their names and their abbreviations in R. Note that the different distributions can be fitted using their R abbreviations (and optionally excluding the brackets) i.e. family=BI(), family=BI are equivalent.

Distributions 	 R names 	 No of parameters

Beta 	 BE() 	 2

Beta Binomial 	 BB() 	 2

Beta one inflated 	 BEOI() 	 3

Beta zero inflated 	 BEZI() 	 3

Beta inflated 	 BEINF() 	 4

Binomial 	 BI() 	 1

Box-Cox Cole and Green 	 BCCG() 	 3

Box-Cox Power Exponential 	 BCPE() 	 4

Box-Cox-t 	 BCT() 	 4

Delaport 	 DEL() 	 3

Exponential 	 EXP() 	 1

Exponential Gaussian 	 exGAUS() 	 3

Exponential generalized Beta type 2 	 EGB2() 	 4

Gamma 	 GA() 	 2

Generalized Beta type 1 	 GB1() 	 4

Generalized Beta type 2 	 GB2() 	 4

Generalized Gamma 	 GG() 	 3

Generalized Inverse Gaussian 	 GIG() 	 3

Generalized t 	 GT() 	 4

Gumbel 	 GU() 	 2

Inverse Gaussian 	 IG() 	 2

Johnson's SU 	 JSU() 	 4

Logistic 	 LO() 	 2

log-Normal 	 LOGNO() 	 2

log-Normal (Box-Cox) 	 LNO() 	 3 (1 fixed)

Negative Binomial type I 	 NBI() 	 2

Negative Binomial type II 	 NBII() 	 2

Normal Exponential t 	 NET() 	 4 (2 fixed)

Normal 	 NO() 	 2

Normal Family 	 NOF() 	 3 (1 fixed)

Power Exponential 	 PE() 	 3

Power Exponential type 2 	 PE2() 	 3

Poison 	 PO() 	 1

Poisson inverse Gaussian 	 PIG() 	 2

Reverse generalized extreme 	 RGE() 	 3

Reverse Gumbel 	 RG() 	 2

Skew Power Exponential type 1 	 SEP1() 	 4

Skew Power Exponential type 2 	 SEP2() 	 4

Skew Power Exponential type 3 	 SEP3() 	 4

Skew Power Exponential type 4 	 SEP4() 	 4

Shash 	 SHASH() 	 4

Sichel (original) 	 SI() 	 3

Sichel (mu as the maen) 	 SICHEL() 	 3

Skew t type 1 	 ST1() 	 3

Skew t type 2 	 ST2() 	 3

Skew t type 3 	 ST3() 	 3

Skew t type 4 	 ST4() 	 3

Skew t type 5 	 ST5() 	 3

t-distribution 	 TF() 	 3

Weibull 	 WEI() 	 2

Weibull(PH parameterization) 	 WEI2() 	 2

Weibull (mu as mean) 	 WEI3() 	 2

Zero inflated poisson 	 ZIP() 	 2

Zero inf. poiss.(mu as mean) 	 ZIP2() 	 2

Zero adjusted IG 	 ZAIG() 	 2

Note that some of the distributions are in the package gamlss.dist. The parameters of the distributions are in order, mu for location, sigma for scale (or dispersion), and nu and tau for shape. More specifically for the BCCG family mu is the median, sigma approximately the coefficient of variation, and nu the skewness parameter. The parameters for BCPE distribution have the same interpretation with the extra fourth parameter tau modelling the kurtosis of the distribution. The parameters for BCT have the same interpretation except that sigma*((tau/(tau-2))^0.5) is approximately the coefficient of variation.

All of the distribution in the above list are also provided with the corresponding d, p, q and r functions for density (pdf), distribution function (cdf), quantile function and random generation function respectively, (see individual distribution for details).

Value

The above GAMLSS families return an object which is of type gamlss.family. This object is used to define the family in the gamlss() fit.

Note

More distributions will be documented in later GAMLSS releases. Further user defined distributions can be incorporate relatively easy, see, for example, the help documentation accompanying the gamlss library.

Author(s)

Mikis Stasinopoulos d.stasinopoulos@londonmet.ac.uk, Bob Rigby r.rigby@londonmet.ac.uk and Calliope Akantziliotou

References

Rigby, R. A. and Stasinopoulos D. M. (2005). Generalized additive models for location, scale and shape,(with discussion), Appl. Statist., 54, part 3, pp 507-554.

Stasinopoulos D. M., Rigby R.A. and Akantziliotou C. (2006) Instructions on how to use the GAMLSS package in R. Accompanying documentation in the current GAMLSS help files, (see also http://www.gamlss.com/).

Stasinopoulos D. M. Rigby R.A. (2007) Generalized additive models for location scale and shape (GAMLSS) in R. Journal of Statistical Software, Vol. 23, Issue 7, Dec 2007, http://www.jstatsoft.org/v23/i07.

See Also

gamlss,BE,BB,BEINF,BI,LNO,BCT,BCPE,BCCG, GA,GU,JSU,IG,LO,NBI,NBII,NO,PE ,PO,RG,PIG,TF,WEI,WEI2,ZIP

Examples


 normal<-NO(mu.link="log", sigma.link="log")
 normal

[Package gamlss.dist version 3.0-0 Index]